diabetes_project/requirements.txt
django==4.2.11
djangorestframework==3.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
minio==7.2.0
boto3==1.28.17
django-storages==1.14.2
drf-spectacular==0.27.0
uvicorn==0.23.2
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0
openai==0.28.1
django-cors-headers==4.3.1
whitenoise==6.6.0
python-json-logger==2.0.7